Move HTTP server definitions in a separate file
